more work on backend refactor

+++ server/utils/FileUtils.js
@@ -2,22 +2,24 @@ import fs from "fs";
 import path from "path";
 
 
-export function emptyDirectory(path){
+export function emptyDirectory(path) {
 
-  if(!directoryExists(path)){
+  if (!directoryExists(path)) {
     throw new Error("Can't empty directory. Directory does not exist");
-  } 
-  
-  
+  }
+
+
   fs.readdir(path, (err, files) => {
     if (err) {
       throw new Error(err);
     }
 
     for (const file of files) {
-      fs.unlink(path + file, (err) => {
-        if (err) console.log(err);
-      });
+      if (fs.lstatSync(path + file).isFile()) {
+        fs.unlink(path + file, (err) => {
+          if (err) console.log(err);
+        });
+      }
     }
   });
 };
@@ -25,16 +27,16 @@ export function emptyDirectory(path){
 
 
 
-export function directoryExists(path){
+export function directoryExists(path) {
 
-  if (fs.existsSync(path)){
+  if (fs.existsSync(path)) {
     return true;
   }
   return false;
 }
-export function makeDirectory(path){
+export function makeDirectory(path) {
 
-  fs.mkdirSync(path, {recursive:true});
+  fs.mkdirSync(path, { recursive: true });
 
   return true;
 }
